Removes/locks behind SRM remaining tribal weapon crafting recipes. (#4581)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I kept the Sword, the Axe and the Spear in #3296, because there were no
other melee weapons available. I think the machettes and combat knives
that are now in cargo satisfy it, so I am:
- Removing the bone sword.
- Locking the bone spear and the bone axe behind being an SRM role
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Changelog

:cl:
del: Bone axe and bone sword are no longer craftable.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->