Coffee 2 (#4451)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- [x] Reflavor coffee cartridges and models
- [x] Make sure everything works properly
- [ ] Add to ships(?)

This request does the following:
* Ports https://github.com/BeeStation/NSV13/pull/2119
  * Which by extension ports:
    * https://github.com/tgstation/tgstation/pull/68178
    * https://github.com/tgstation/tgstation/pull/69505
    * https://github.com/tgstation/tgstation/pull/69347
    * Partially https://github.com/tgstation/tgstation/pull/69386
    * https://github.com/tgstation/tgstation/pull/70991
 

![imagen](https://github.com/user-attachments/assets/ee4f75f7-b6ed-4555-9efe-34ba1a659f08)


![imagen](https://github.com/user-attachments/assets/beb6a63f-fb84-4bd7-a81d-012e8b9101da)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Coffee machines, and two different qualities of them, are flavorful

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Coffee machines, coffeepots, coffee cups, and other coffee stuff
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->