Adds a mission to the Abandoned Listening Post (#4614)
## About The Pull Request

Adds a single documents retrieval mission to the Abandoned Listening
Post, to better thematically fit the ruin.

I was a little frustrated that my ruin only had a generic blackbox
mission, so I came up with something that fit what the ruin is supposed
to be.

## Why It's Good For The Game

Proper missions housekeeping on one of my ruins is probably a good
thing.

## Changelog

:cl:
add: New Dynamic Mission for the Abandoned Listening Post
/:cl:

---------

Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>