End-Of-Round Bluespace Jumps only affect ships already in transit (#4681)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Makes end of round bsj's only affect ships that are in transit, ignoring
ships that are docked. Because end of round bsj's effectively only enter
ships into transit, this means that ships are now unaffected by the end
of round jumps. However, if there are any other effects that should be
triggered by jumping they still will be for ships in transit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Doesn't leave behind crew if they're late for the jump, and gives people
a bit more time before the end of the round without needing to rush back
to the ship.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: end of round BSJ no longer affects docked ships
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->