Adds (placebo) prescription pills to general loadout (#4683)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a bottle of prescription pills and a prescription autoinjector to
loadout. They don't do anything aside from putting a message in your
chat saying you feel better, but you can use them for flavor if you'd
like.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This was asked as a low-impact solution to some players wanting to
roleplay their blorbos taking medicine.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Prescription pills added to loadout! Filled with placebatol, they
do almost nothing, but are good for flavoring your character.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Rug-Dealer <switchblade2011@gmail.com>
Co-authored-by: thgvr <81882910+thgvr@users.noreply.github.com>