Adds oxygen candles (#4667)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds oxygen candles, ported from Monke, who ported them from Nova. One
pull of this portable device will produce profuse plumes of oxygen for
pulmonary purposes.
You can find one stocked in an emergency toolbox, or they can be bought
for 50 credits each.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Too often are we faced with circumstances where a room has not quite
enough oxygen to be comfortable to breathe, and the nature of the airmix
distribution makes it hard to replenish these rooms without dragging a
canister or a pump in to re-oxygenate the room.
Oxygen candles present a portable -- but consumable! -- way to replenish
oxygen back to comfortable breathing levels. Just don't use them in very
enclosed spaces.

I guess you could bring these things to ruins too if you need a pocket
of breathable air (maybe you need to eat or do CPR?).
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Oxygen candles! Stocked in emergency (red) toolboxes or purchasable
from the outpost; simply pull the cord and place to quickly re-oxygenate
rooms.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Rug-Dealer <switchblade2011@gmail.com>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>