Dismantle Light Switches + Mount Fix (#4750)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Ports https://github.com/PentestSS13/Pentest/pull/278

Light switches could not be dismantled from what I've tried, and could
only be destroyed. This corrects that. Screwdriver and crowbar can both
be used to create a frame after a short delay. Additionally, I have
fixed the pixel offset and direction when mounting light switches on the
wall.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It hurts my soul to smack light switches until they die.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Light switches can now be properly dismantled into frames with
crowbar or screwdriver.
fix: Light switches are mounted with the right directional set and with
the correct offset.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: FslashN <150570209+FslashN@users.noreply.github.com>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>