Adds buckles to toilets (#4714)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Lets you buckle to a toilet like you would a chair. That's about it.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Being vaguely chair-shaped and nominally rated for microgravity
operations, the toilet _should_ have a set of buckles for securing
yourself to the throne in-case you need to use it while the gravity is
out, and more practically as a way of preventing yourself from
embarrassingly falling over during takeoff.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Following a review of shipbuilding standards, all toilets are now
equipped with the same set of buckles that chairs have.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->